requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
feedparser>=6.0.0
psutil>=5.9.0
python-dateutil>=2.8.0
apscheduler>=3.10.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
flake8>=6.0.0